requirements.txt
Sphinx
pip-tools
pip>=1.4
pylint
python-dateutil==1.5
setuptools>=0.9
ujson
urllib3
lxml
wheel>=0.21
gunicorn
flask
python-telegram-bot
psycopg2
Flask-SQLAlchemy
Flask-Migrate